requirements.txt
Django==1.5.1
Pillow==2.0.0
argparse==1.2.1
wsgiref==0.1.2
